Literacy
Literacy is more difficult to get as the education access per wealth is 25% of what it was before (0.005 instead of 0.02 in vanilla and previous versions). This means that universities are even more important than they were previously. (If you had 10 wealth, you'd have 20% education access in the state). 15 base innovation now as well. More tech from spread.